+++ sport-log-types/src/types/diary_wod.rs
@@ -29,7 +29,7 @@ use crate::{
     )
 )]
 #[cfg_attr(feature = "full", sql_type = "diesel::sql_types::Integer")]
-pub struct DiaryId(pub i32);
+pub struct DiaryId(pub i64);
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[cfg_attr(
@@ -87,7 +87,7 @@ pub struct NewDiary {
     )
 )]
 #[cfg_attr(feature = "full", sql_type = "diesel::sql_types::Integer")]
-pub struct WodId(pub i32);
+pub struct WodId(pub i64);
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[cfg_attr(
